Rename utils to lib.
